suppression des signatures de methodes inutiles dans Patientrepository, suppression de la fonctionnalité deleteAll, refactoring du nom du projet en 'patient'
