+++ patient/pom.xml
@@ -12,7 +12,7 @@
 	<groupId>com.mediscreen</groupId>
 	<artifactId>PreventDiabetes</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>PreventDiabetes</name>
+	<name>Patient</name>
 	<description>Spring Boot project for clinique Abernathy</description>
 	<properties>
 		<java.version>11</java.version>
