+++ patient/src/main/java/com/mediscreen/patient/PatientApplication.java
@@ -1,0 +1,13 @@
+package com.mediscreen.patient;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PatientApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PatientApplication.class, args);
+	}
+
+}
