+++ patient/src/main/java/com/mediscreen/patient/model/Patient.java
@@ -1,16 +1,12 @@
-package com.mediscreen.preventdiabetes.model;
+package com.mediscreen.patient.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import java.lang.module.FindException;
-import java.math.BigInteger;
 
 @AllArgsConstructor
 @NoArgsConstructor
