+++ patient/src/main/java/com/mediscreen/patient/repository/PatientRepository.java
@@ -1,0 +1,12 @@
+package com.mediscreen.patient.repository;
+
+import com.mediscreen.patient.model.Patient;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface PatientRepository extends MongoRepository<Patient, String> {
+    Optional<Patient> findByLastNameAndFirstName(String lastName, String firstName);
+}
