+++ patient/src/main/java/com/mediscreen/patient/utils/LightPatient.java
@@ -1,4 +1,4 @@
-package com.mediscreen.preventdiabetes.utils;
+package com.mediscreen.patient.utils;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
