--- patient/src/main/java/com/mediscreen/preventdiabetes/PreventDiabetesApplication.java
@@ -1,14 +1,0 @@
-package com.mediscreen.preventdiabetes;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-
-@SpringBootApplication
-public class PreventDiabetesApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PreventDiabetesApplication.class, args);
-	}
-
-}
