+++ patient/src/main/resources/application.properties
@@ -1,27 +1,3 @@
 spring.application.name=PreventDiabetes
 server.port=8081
 spring.data.mongodb.uri=mongodb+srv://fercakflorian:1234@cluster.hqp8whj.mongodb.net/PreventDiabetes.patient
-#spring.data.mongodb.auto-index-creation=true
-#spring.data.mongodb.authentication-database=fercakflorian
-#spring.data.mongodb.password=1234
-#spring.data.mongodb.database=PreventDiabetes
-#logging.level.org.springframework.boot.web.embedded.tomcat=INFO
-#
-#
-#logging.level.org.springframework=INFO
-#
-#################### DataSource Configuration ##########################
-##spring.datasource.url=jdbc:mysql://localhost:3306/prevent_diabetes?serverTimezone=UTC
-##spring.datasource.username=root
-##spring.datasource.password=rootroot
-##spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-##spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-#
-#
-#################### Hibernate Configuration ##########################
-#spring.sql.init.mode = always
-#
-#spring.jpa.hibernate.ddl-auto = none
-##spring.jpa.hibernate.ddl-auto=update
-#
-##spring.jpa.show-sql=true
