+++ patient/src/test/java/com/mediscreen/patient/PatientApplicationTests.java
@@ -1,10 +1,10 @@
-package com.mediscreen.preventdiabetes;
+package com.mediscreen.patient;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PreventDiabetesApplicationTests {
+class PatientApplicationTests {
 
 	@Test
 	void contextLoads() {
